moved theme api from core to canvas
